+++ integration-tests/src/test/java/it/matcher/ParamMatchers.java
@@ -2,16 +2,19 @@ package it.matcher;
 
 import org.hamcrest.Matcher;
 
+import java.util.TimeZone;
+
 import static it.matcher.MatchesPattern.matchesPattern;
+import static org.hamcrest.Matchers.isIn;
 
 /**
  * Matchers for parameters passed to Connect add-ons.
  */
-public class ParamMatchers
-{
+public class ParamMatchers {
+
     public static Matcher<String> isTimeZone()
     {
-        return matchesPattern("[A-Za-z0-9_\\-]+/[A-Za-z0-9_\\-]+");
+        return isIn(TimeZone.getAvailableIDs());
     }
 
     public static Matcher<String> isLocale()