+++ plugin/src/main/java/com/atlassian/plugin/connect/plugin/capabilities/module/macro/StaticContentMacro.java
@@ -1,6 +1,8 @@
 package com.atlassian.plugin.connect.plugin.capabilities.module.macro;
 
 import com.atlassian.confluence.content.render.xhtml.ConversionContext;
+import com.atlassian.confluence.content.render.xhtml.macro.annotation.Format;
+import com.atlassian.confluence.content.render.xhtml.macro.annotation.RequiresFormat;
 import com.atlassian.confluence.macro.MacroExecutionException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -24,6 +26,7 @@ public class StaticContentMacro extends AbstractMacro
     }
 
     @Override
+    @RequiresFormat(Format.Storage)
     public String execute(Map<String, String> parameters, String body, ConversionContext context) throws MacroExecutionException
     {
         return remoteMacroRenderer.executeStatic(addOnKey, moduleKey, uriTemplate, parameters, body, context);