+++ plugin/src/main/java/com/atlassian/plugin/connect/plugin/usermanagement/ConnectAddOnUserUtil.java
@@ -1,13 +1,76 @@
 package com.atlassian.plugin.connect.plugin.usermanagement;
 
+import com.atlassian.crowd.exception.ApplicationNotFoundException;
+import com.atlassian.crowd.model.user.User;
+import com.atlassian.crowd.service.client.ClientProperties;
+import com.atlassian.crowd.service.client.ClientPropertiesImpl;
+import com.atlassian.crowd.service.client.ClientResourceLocator;
+import com.google.common.collect.ImmutableMap;
+
+import java.util.Collections;
+import java.util.Set;
+
 import static com.google.common.base.Preconditions.checkNotNull;
 
 public class ConnectAddOnUserUtil
 {
-    public static String  usernameForAddon(String addonKey)
+    public static String usernameForAddon(String addonKey)
     {
         checkNotNull(addonKey);
         return Constants.ADDON_USERNAME_PREFIX + addonKey;
+    }
+
+    /**
+     * Will return the a Map containing a value that will mark the Connect AddOn user as a 'Connect User'. This will be a 'synch-able'
+     * Crowd attribute (shared attribute across all Crowd connected applications) when this is finally enabled in Crowd.
+     * @param applicationName the name of the application the attribute was created from. This is mostly for an audit trail.
+     * @return An ImmutableMap allowing the marking of a user as a Connect AddOn user
+     * @throws com.atlassian.crowd.exception.ApplicationNotFoundException
+     */
+    public static ImmutableMap<String, Set<String>> buildConnectAddOnUserAttribute(String applicationName) throws ApplicationNotFoundException
+    {
+        return ImmutableMap.of(buildAttributeConnectAddOnAttributeName(applicationName), Collections.singleton("true"));
+    }
+
+    /**
+     * Builds the Connect AddOn attribute name that is used to store against Connect AddOn Users, which 'marks' them as a Connect AddOn.
+     * @param applicationName the name of the application the attribute was created from. This is mostly for an audit trail.
+     * @return String representation of the Connect AddOn User attribute name. <em>synch.APPLICATION_NAME.atlassian-connect-user</em>
+     */
+    public static String buildAttributeConnectAddOnAttributeName(String applicationName)
+    {
+        return "synch." + applicationName + ".atlassian-connect-user";
+    }
+
+    /**
+     * Will validate that a user's username adheres to the correct AddOn format
+     * @param user the user to validate
+     * @return true if the user has a valid username
+     */
+    public static boolean validAddOnUsername(User user)
+    {
+        String name = user.getName();
+        return name != null && name.startsWith(Constants.ADDON_USERNAME_PREFIX);
+    }
+
+    /**
+     * Will validate that the user has an email address that equals the desired AddOn email ({@link Constants#ADDON_USER_EMAIL_ADDRESS}).
+     * @param user the user to validate
+     * @return true if the email address is valid
+     */
+    public static boolean validAddOnEmailAddress(User user)
+    {
+        return Constants.ADDON_USER_EMAIL_ADDRESS.equals(user.getEmailAddress());
+    }
+
+    /**
+     * Will return the required configuration to communicate with the configured RemoteCrowd Server
+     * @return authentication and configuration properties for the RemoteCrowd Server
+     */
+    public static ClientProperties getClientProperties()
+    {
+        ClientResourceLocator resourceLocator = new ClientResourceLocator(com.atlassian.crowd.integration.Constants.PROPERTIES_FILE);
+        return ClientPropertiesImpl.newInstanceFromResourceLocator(resourceLocator);
     }
 
     public static class Constants